changed data types from tinyint to int to perhaps cater for the future?

+++ Database.AnimalFiveHead/dbo/Tables/dbo.tb_PlayerSessionInformation.sql
@@ -1,8 +1,8 @@
 CREATE TABLE [dbo].[tb_PlayerSessionInformation]
 (
     [SessionId] UNIQUEIDENTIFIER NOT NULL, 
-    [PlayerId] TINYINT NOT NULL, 
-    [Score] SMALLINT NOT NULL, 
+    [PlayerId] INT NOT NULL, 
+    [Score] INT NOT NULL, 
     [Cards] NVARCHAR(MAX) NOT NULL, 
     [CardIds] NVARCHAR(MAX) NOT NULL, 
     [DateTimeAdded] DATETIME2 NULL,
